Harden release governance and Vercel sandbox billing
