+++ .env.example
@@ -142,8 +142,23 @@ OVERLAY_FEATURE_REMOTE_AGENT_RUNS
 # Deferred: keep false until the managed-host image, live conformance, billing,
 # cleanup, and security release gates are all approved.
 OVERLAY_FEATURE_OVERLAY_CLOUD_ENVIRONMENTS
-OVERLAY_MANAGED_SANDBOX_PROVIDER
+# Overlay Cloud defaults to Vercel Sandbox. Daytona is opt-in only.
+OVERLAY_MANAGED_SANDBOX_PROVIDER=vercel
 OVERLAY_AGENT_HOST_IMAGE
+# Pin the default region so the configured unit prices match the provider bill.
+OVERLAY_VERCEL_SANDBOX_REGION=iad1
+# Current Vercel Sandbox on-demand rates. Override these together if the
+# selected region or Vercel contract uses different prices.
+OVERLAY_VERCEL_SANDBOX_ACTIVE_CPU_USD_PER_HOUR=0.128
+OVERLAY_VERCEL_SANDBOX_MEMORY_USD_PER_GB_HOUR=0.0212
+OVERLAY_VERCEL_SANDBOX_EGRESS_USD_PER_GB=0.15
+OVERLAY_VERCEL_SANDBOX_CREATION_USD=0.0000006
+# Reserve above the estimated maximum and reject runs above the provider cap.
+OVERLAY_VERCEL_SANDBOX_RESERVATION_BUFFER_PERCENT=25
+OVERLAY_SANDBOX_MAX_PROVIDER_COST_USD_PER_RUN=15
+OVERLAY_SANDBOX_PROVIDER_SPEND_ALERT_USD=10
+# Standalone /api/v1/sandbox/run reserves for up to 10 decimal GB of egress.
+OVERLAY_SANDBOX_RUN_MAX_EGRESS_BYTES=10000000000
 VERCEL_TOKEN
 VERCEL_TEAM_ID
 VERCEL_PROJECT_ID
