+++ .github/CODEOWNERS
@@ -1,0 +1,2 @@
+# Default review owner for product, infrastructure, billing, and security changes.
+* @DevelopedByDev
