+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -1,0 +1,37 @@
+name: Bug report
+description: Report a reproducible Overlay problem
+title: "[Bug]: "
+labels: [bug]
+body:
+  - type: markdown
+    attributes:
+      value: Do not include credentials, customer data, or security vulnerabilities. Report vulnerabilities through SECURITY.md.
+  - type: textarea
+    id: description
+    attributes:
+      label: What happened?
+      description: Describe the observed and expected behavior.
+    validations:
+      required: true
+  - type: textarea
+    id: reproduction
+    attributes:
+      label: Reproduction
+      description: Provide the smallest reliable sequence of steps.
+    validations:
+      required: true
+  - type: input
+    id: version
+    attributes:
+      label: Version or commit
+      placeholder: Commit SHA, deployment, or app version
+  - type: textarea
+    id: environment
+    attributes:
+      label: Environment
+      description: Browser, operating system, deployment type, and relevant non-secret configuration.
+  - type: textarea
+    id: logs
+    attributes:
+      label: Redacted logs
+      render: shell
