+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,0 +1,5 @@
+blank_issues_enabled: false
+contact_links:
+  - name: Security vulnerability
+    url: https://github.com/LayerNorm/overlay-web/security/advisories/new
+    about: Privately report a security vulnerability.
