+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,18 @@
+## What changed
+
+Describe the outcome and the affected surfaces.
+
+## Risk and rollback
+
+- Security or tenant-boundary impact:
+- Billing or provider-cost impact:
+- Data migration impact:
+- Rollback plan:
+
+## Verification
+
+- [ ] Relevant focused tests pass
+- [ ] Typecheck passes when contracts changed
+- [ ] Living documentation was updated when required
+- [ ] No credentials or customer data are included
+- [ ] Browser or deployment evidence is attached when behavior depends on a hosted environment
