+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,11 @@
+# Code of Conduct
+
+Overlay contributors must help maintain a respectful, harassment-free community for everyone, regardless of identity, background, experience, or viewpoint.
+
+Examples of expected behavior include constructive feedback, respect for differing perspectives, accountability for mistakes, and attention to the community's safety. Harassment, threats, discriminatory conduct, unwanted sexual attention, doxxing, and sustained disruption are unacceptable.
+
+Project maintainers may remove, edit, or reject contributions and may temporarily or permanently restrict participation when conduct is inappropriate, threatening, or harmful.
+
+Report conduct concerns privately to `divyansh@layernorm.co`. Reports will be reviewed as promptly and confidentially as practical. Do not use public issues for sensitive reports.
+
+This policy is adapted from the [Contributor Covenant, version 2.1](https://www.contributor-covenant.org/version/2/1/code_of_conduct.html).
