+++ config/on-prem-convex-runtime-baseline.json
@@ -167,6 +167,7 @@
     "src/app/api/v1/projects/share-directory/route.ts",
     "src/app/api/v1/providers/connections/route.ts",
     "src/app/api/v1/providers/connections/test/route.ts",
+    "src/app/api/v1/sandbox/run/route.ts",
     "src/app/api/v1/search/route.ts",
     "src/app/api/v1/settings/route.ts",
     "src/app/api/v1/shares/impact/route.ts",
